Make MAC addresses build-time configurable
The Makefile will now generate a random MAC address range to use for
all components.  Individual components add an offset to the base MAC address

The Makefile saves the MAC address it uses and reuses the same one
if it finds a saved one, so rerunning make does not change the MAC address
unless you delete the build tree.

Signed-off-by: Peter Chubb <peter.chubb@unsw.edu.au>